src/backend/requirements.txt
fastapi
uvicorn
openai
anthropic
requests
python-dotenv
chromadb
google-genai
neo4j
google-generativeai
networkx
numpy
voyageai